Simplify .bash_profile to only include .bashrc & .profile
tmux only reads .bash_profile [which is a stupid behavior], so, if you
defined your aliases/path in bashrc, it never loaded in tmux. If however
you defined them in bash_profile, you would have lost them in a normal
shell, since .bash_profile is only meant to be run for an interactive
login shell.
To address the problem, now .bashrc holds all the bash functions and
aliases and .profile holds all the PATH exports and .bash_profile only
includes both of them ;)